requirements.front.txt
pandas
numpy

streamlit
requests

pytest
pytest-flask
flake8
black
pylint

Ipython